Moving tests to individual processes. Coverage is broke and output is more difficult to read - ideas welcome.
